+++ app/serializers/admin_journal_role_serializer.rb
@@ -1,0 +1,6 @@
+# Provides a display-only representation of a role
+class AdminJournalRoleSerializer < ActiveModel::Serializer
+  attributes :id,
+             :name,
+             :journal_id
+end