+++ db/migrate/20160803202527_add_hint_column_to_roles.rb
@@ -1,0 +1,6 @@
+class AddHintColumnToRoles < ActiveRecord::Migration
+  def change
+    add_column :roles, :assigned_to_type_hint, :string
+    add_index :roles, :assigned_to_type_hint
+  end
+end