+++ spec/models/paper_spec.rb
@@ -643,6 +643,14 @@ describe Paper do
         paper.withdraw! "Some reason", withdrawn_by_user
         expect(paper).to_not be_editable
       end
+
+      it "broadcasts 'paper:withdrawn' event" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:withdrawn")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.withdraw! 'reason', withdrawn_by_user
+      end
     end
 
     describe '#invite_full_submission' do
@@ -792,6 +800,14 @@ describe Paper do
           paper.reject!
           expect(paper.rejected?).to be true
         end
+
+        it "broadcasts 'paper:rejected' event" do
+          allow(Notifier).to receive(:notify)
+          expect(Notifier).to receive(:notify).with(hash_including(event: "paper:rejected")) do |args|
+            expect(args[:data][:record]).to eq(paper)
+          end
+          paper.reject!
+        end
       end
 
       context 'paper is initially_submitted' do
@@ -912,12 +928,36 @@ describe Paper do
           expect(paper.accepted_at.utc).to be_within_db_precision.of now
         end
       end
+
+      it "broadcasts 'paper:accepted' event" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:accepted")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.accept!
+      end
+    end
+
+    context "rejection" do
+      it 'rejects the paper' do
+        decision = instance_double('Decision', verdict: 'reject')
+        expect(paper).to receive(:reject!)
+        paper.make_decision decision
+      end
     end
 
     describe "#major_revision!" do
       it "puts the paper in_revision" do
         paper.major_revision!
         expect(paper.publishing_state).to eq("in_revision")
+      end
+
+      it "broadcasts 'paper:in_revision'" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:in_revision")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.major_revision!
       end
     end