+++ spec/models/paper_spec.rb
@@ -595,6 +595,14 @@ describe Paper do
         paper.withdraw! "Some reason", withdrawn_by_user
         expect(paper).to_not be_editable
       end
+
+      it "broadcasts 'paper:withdrawn' event" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:withdrawn")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.withdraw! 'reason', withdrawn_by_user
+      end
     end
 
     describe '#invite_full_submission' do
@@ -728,6 +736,14 @@ describe Paper do
           paper.reject!
           expect(paper.rejected?).to be true
         end
+
+        it "broadcasts 'paper:rejected' event" do
+          allow(Notifier).to receive(:notify)
+          expect(Notifier).to receive(:notify).with(hash_including(event: "paper:rejected")) do |args|
+            expect(args[:data][:record]).to eq(paper)
+          end
+          paper.reject!
+        end
       end
 
       context 'paper is initially_submitted' do
@@ -783,11 +799,22 @@ describe Paper do
         paper.make_decision decision
         expect(paper.accepted_at.utc).to be_within(1.second).of Time.zone.now
       end
+
+      it "broadcasts 'paper:accepted' event" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:accepted")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.make_decision decision
+      end
     end
 
     context "rejection" do
+      let(:decision) do
+        instance_double('Decision', verdict: 'reject')
+      end
+
       it 'rejects the paper' do
-        decision = instance_double('Decision', verdict: 'reject')
         expect(paper).to receive(:reject!)
         paper.make_decision decision
       end
@@ -809,6 +836,14 @@ describe Paper do
         paper.make_decision decision
         expect(paper.latest_version.major_version).to be(1)
         expect(paper.latest_version.minor_version).to be(0)
+      end
+
+      it "broadcasts 'paper:in_revision'" do
+        allow(Notifier).to receive(:notify)
+        expect(Notifier).to receive(:notify).with(hash_including(event: "paper:in_revision")) do |args|
+          expect(args[:data][:record]).to eq(paper)
+        end
+        paper.make_decision decision
       end
     end