+++ CHANGELOG.md
@@ -5,6 +5,8 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ## [Version] - {Date}
 ### Added
 - Daily export of billing log
+- Admins can view journal roles assigned to users
+- AE country is included in Apex export
 ### Changed
 ### Deprecated
 ### Removed
@@ -32,7 +34,7 @@ editability down the line.
 - Require only a single click to register a decision
 - Reviewer recommendations display full reason in accordion and workflow view.
 - Attachments ( Figures, SI files, etc. ) will be available without downtime during deploy.
-
+- Fixed comparing versions to highlight sentence changes instead of entire document
 ### Security
 
 ## [1.4.18] - {2016-07-06}