+++ CHANGELOG.md
@@ -9,6 +9,8 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 - AE country is included in Apex export
 - Discussion Topic title validation
 ### Changed
+- Soft deleting questions and answers. Answers are retained even when questions are marked
+  as deleted
 ### Deprecated
 ### Removed
 ### Fixed